refactor: extract AttributionItem.standard and SettingsRowStyling for reuse by anonymous Settings (tc-hq9h7.3)
